commit: Use structured errors in `GetTreeEntries` rpc
In the `GetTreeEntries` rpc use structured errors for the following
scenarios:
1. the provided path is invalid
2. the provided revision is invalid

We add the usage of structured errors behind a featureflag
`GetTreeEntriesStructuredErrors`, which was introduced in a previous
commit. This allows for safe rollout. We also amend the tests and add
some new tests to cover more scenarios.

Some errors previously just returned an empty response. Rails has now
been patched to expect errors and we now add structured errors in such
scenario's.