stats: Expose absolute and relative paths for alternates
Currently the Git alternates file parsing done with
`stats.AlternatesInfoForRepository()` always converts the alternates
directory paths to be absolute. This is problematic because sometimes it
is required to see what the exact path of the alternates recorded in the
alternates file is.

This change updates `stats.AlternatesInfoForRepository()` to instead
return the raw directory paths contained within a repository's Git
alternates file. `AbsoluteObjectDirectories()` has been added to convert
to absolute paths if needed.