praefect: Add structerr interceptors to Praefect binary test build
While the preceding commit has set up structerr interceptors for the
in-process gRPC server, the actual Praefect binary will not use them.
While this is what we want for production, our `test-with-praefect` test
target runs Praefect as a separate process by executing its binary. This
part of our testing infrastructure thus doesn't yet have interceptors
installed.

The Praefect binary is built via `testcfg.BuildPraefect()`, which always
sets up the `gitaly_test` build tag. We can thus piggy-back on this tag
and conditionally compile default server options depending on whether
the tag is set or not. If it isn't we leave the options empty, but if it
is set then we set up our structerr interceptors.