storage: Convert `VerifyRepository()` to accept a `Repository`
As `VerifyRepository()` was implemented as a standalone function it by
necessity had to accept a full path as input, as it otherwise wouldn't
have been able to compute the actual repository path if it only got the
`storage.Repository` as input. Now that we have converted it to be a
receiver function though we can also have it accept the repository
itself.

Refactor the function to accept a `storage.Repository` and adust callers
accordingly. This will allow us to move all verification of the repo
from `GetRepoPath()` into `VerifyRepository()` to have better separation
of concerns.