objectpool: Simplify computation of object pool paths
The `FromRepo()` function constructs an object pool from an object pool
member by reading the alternates file and computing the corresponding
object pool path from it. The logic is quite hard to follow though as it
is needlessly split up across two different functions.

Simplify the logic by inlining the function call. This change also
prepares us for the upcoming refactoring of `ValidateRepository()`,
which is about to be converted to accept a `storage.Repository` instead
of the repository path.