internalgitaly: Prepare repo walker for ValidateRepository changes
We're about to make `ValidateRepository()` a function on the
`storage.Locator` interface that accepts a `storage.Repository` as
argument instead of the repository path. Refactor the repository walker
to for this change by making both the locator and storage name inside of
`walkStorage()`.

Note that this refactoring changes the error code from NotFound to
InvalidArgument when the storage name cannot be found. This is in fact
the right thing to do, as a request for a storage name that is not
configured is invalid. This is in contrast to a request for a storage
name that does exist, but where the storage itself cannot be found,
where NotFound would be the correct error code.