storage: Make `VerifyRepository()` a receiver function
The `VerifyRepository()` function receives the full path to the
repository in order to verify that it is valid. This is kind of an
overreach into storage-specific details that only implementations of the
interface should actually know.

But more importantly, the current design keeps us from providing better
error reporting. The logic to verify repos in the `config.Locator` is
split across `VerifyRepository()` and `GetRepoPath()` as a necessity
because `GetRepoPath()` gets a `storage.Repository` as input while
`VerifyRepository()` gets the repostiroy path. Thus only `GetRepoPath()`
is in a position to verify the request itself.

Refactor the code to make `VerifyRepository()` a receiver function on
the `storage.Locator` interface. This will allow us to refactor it to
also get a `storage.Repository` as input and thus move all validations
of the actual request into `VerifyRepository()` itself.