operation: Refactor UserMergeToRef with helper
Refactor UserMergeToRef's code to put the git2go implementation into a
function. This makes it easier to add the pure git implementation in a
future commit.

We can also get rid of the allow conflicts parameter since we no longer
use it on the Rails side. See https://gitlab.com/gitlab-org/gitlab/-/merge_requests/124239
for the MR that removed the usage of the parameter altogether.