config/locator: Use ErrRepositoryPathNotSet errors
While Praefect uses `ErrRepositoryPathNotSet` in the case where the
passed repository's relative path is not set, Gitaly's `config.Locator`
retuns a generic errror, only. This causes Praefect and Gitaly to have
different error messages for the same condition, which is inconsistent.

Convert the locator to return `ErrRepositoryPathNotSet and amend tests
to check for `storage.ErrRepositoryPathNotSet`. Rewrite the error
message to ensure that there aren't any tests left that check the old
error message explicitly.