config/locator: Use ErrRepositoryNotSet errors
The `ValidateRepository()` function of the `config.Locator` doesn't
explicitly check whether the passed-in repository is set. And right now
it doesn't need to because all RPCs use `service.ValidateRepository()`
anyway to verify that the repository is indeed set. We're about to
remove this helper in favor of `locator.ValidateRepository()` in a
future patch series, at which point the locator would become exposed to
any unset repositories.

Explicitly handle unset repositories in the `config.Locator` and return
the same error as `service.ValidateRepository()`. Refactor tests to
check for `ErrRepositoryNotSet` explicitly and rename the error message
in order to make sure that there cannot be any tests left that check for
the old error message.