protoregistry: Distinguish error message for missing repository field
In order to correctly route a gRPC call to the correct Gitaly storage in
Praefect, the protoregistry package inspects the initial gRPC message to
extract the field that contains the repository via a custom extension.
If this field cannot be found or if the repository is simply not set,
then the package returns an error indicating this condition.

This error has the same message as `storage.ErrRepositoryNotFound`,
which is in fact by design so that we see the same errors for the same
underlying root cause in both Gitaly and Praefect. It is a layering
violation though to have the `protoregistry` use the same error message
as the storage layer for an error that is rather more specific though.

Refactor the error handling to return a concrete error denoting that the
field cannot be found. Existing callers of the `protoregistry` package
are adjusted to handle these and instead return the storage-specific
error.