config: Move validation of repository into `ValidateRepository()`
The verification of a repository is split across `ValidateRepository()`
and `GetRepoPath()` due to historical reasons. This has the consequence
that `ValidateRepository()` only validates a subset of repository infos
and that error reporting is not as good as it could be.

Refactor the code so that `ValidateRepository()` performs all
validations. As both `ValidateRepository()` and `GetRepoPath()` path
have valid usecases where they are called with nonexistent repositories
we also have to introduce a new option that disables the existence check
for the repository in `ValidateRepository()`.

With this change we're able to replace `service.ValidateRepository()`
with `storage.Locator.ValidateRepository()` in our codebase so that the
request and actual existence of a repository can be checked in a single
call. This should help us to tighten input validation for our RPCs and
help us to return more unified errors in the future.