gitaly-gpg: Binary for signing and verifying gpg
Since we are moving away from using gitaly-git2go and using git's
commit-tree to create commits, we need a way to still support signing
with a signing key file. This is not something that git supports out of
the box since it calls gpg(1), which does not support passing in a key
file for signing.

Instead, we can use git's gpg.program configuration to give it a
different binary that we can pass an environment variable for setting
the signing key path.

https://git-scm.com/docs/git-config#Documentation/git-config.txt-gpgprogram
can be consulted for the specifics of the CLI interface that we need to
support.