Move extractSignature into internal package
extractSignature is useful for testing if a commit has been properly
signed. It was used by git2go, but now that we are about to sign commits
manually in the localrepo package, we will want to test it there to.
Prepare for this by extracting the function into a helper in an internal
package.

Also, extractSignature does not support sha256 since the gpg prefix for
sha256 is different. Fix this as well.

Finally, the oid parameter actually never gets used. Let's remove it as
well.