hook: Refactor repository validation to verify existence
We have recently introduced the `locator.VerifyRepository()` function.
In contrast to `service.VerifyRepository()`, this function is also in a
position to verify that the repository path exists and points to a valid
Git repository.

Refactor the service to use this function for request validation. Note
that the repository-creating RPCs naturally skip the existence check as
the repository is not supposed to exist yet.

Signed-off-by: Patrick Steinhardt <psteinhardt@gitlab.com>