ssh: Fix tests for invalid arguments with Praefect
We have some tests in the ssh package which verify that requests fail as
expected when no stdin was provided. But while Gitaly returns expected
errors, the error condition for Praefect is that the repository cannot
be found. This is wrong, but expected given that neither of them sets up
the repository via `gittest.CreateRepository()`, and thus Praefect would
not be able to locate them.

Fix the tests by working with actually existent repositories.

Signed-off-by: Patrick Steinhardt <psteinhardt@gitlab.com>