storage: Replace `ErrInvalidRepository` with finer-grained errors
The `ErrInvalidRepository` error is returned by the Praefect coordinator
in the case where any of the parameters of the repository are invalid.
This is both rather unspecific as it doesn't pinpoint the root cause and
inconsistent with Gitaly, which returns concrete error conditions for
the different parameters.

Replace the error with the same finer-grained errors that Gitaly uses.