gitaly/hook: Adjust tests to start verifying transactional votes
We're about to start synchronizing hook execution on primary and
secondary nodes. This will be implemented via a transactional vote, but
as we don't yet perform transactional voting in the hooks themselves we
of course aren't checking for them in our tests.

Adjust the tests to check votes which have been cast as a preparatory
step for the upcoming change.