config/locator: Use constructor for ErrRepositoryNotFound
We're not using the constructor `NewRepositoryNotFoundError()` when the
repository that we're validating does not exist, but instead construct
our own error that has the full repository path as error metadata. This
has historical reasons because the function used to not have access to
the `storage.Repository` and was thus couldn't use the constructor that
accepts both the storage name and the relative path.

Refactor the function to use the constructor so that errors are more
consistent.