localrepo: Parse `MergeTree` errors only for exit status 1
The manual page for git-merge-tree(1) states that the exit status is set
to 1, only when there is conflicts. If there are other errors, the exit
status is neither 0 (which is used to denote no conflicts) or 1.

In the `MergeTree` however we default to parsing conflict errors when
the status is not >1. This shouldn't matter as long as `ExecAndWait()`
doesn't return an exist status of 0 plus an error. But is confusing
nevertheless. Let's rewrite the code to explicitly check for exit status
1.