praefect: Introduce `GetObjectPoolHandler()`
The `GetObjectPool()` RPC should return object pool information that can
be consumed by the client. When this Gitaly RPC is executed behind
Praefect, the replica path is returned instead of the relative path. The
replica path is an implementation detail of Praefect and not useful to
the client which really wants the relative path.

This change implements a `GetObjectPoolHandler()` to explicitly handle
the `GetObjectPool()` RPC in Praefect instead of relying on the
`UnknownServiceHandler` to proxy the RPC. This enables the RPC's
response replica path to be reversed by Praefect to the appropriate
relative path and storage the client expects.