backup: Introduce a vanity repository to calculate paths
When we generate a server-side backup, we want to be able to use any
arbitrary replica to then stream the backup directly to object storage.
The problem is that replicas through praefect have rewritten storages
and relative paths. So in order to generate a backup with praefect
paths, we need to have both: the repository used to access on the
filesystem, and the repository used to generate backup paths.