internal service: Remove BackupRepos RPC
The streaming RPC BackupRepos is being replaced with BackupRepository as
it was determined that streaming multiple repositories through praefect
would be too difficult. Instead we'll use a simple single repository
RPC.

BackupRepos is part of the server-side backups work that is not
functional yet. So it is safe to remove the RPC.

This reverts commit 11825b3ad89525b194cc4095e581eef843377cdb, reversing
changes made to 760051f1fcd51c71e5eaa720f4a55a5bc9061a1a.