praefect/coordinator: Untangle NotFound errors from storage layer
The datastore is reusing errors from the storage layer even though these
are conceptually orthogonal to each other. Furthermore, given that we
sometimes don't have the storage name and relative path available but
just the repository ID, we need to use different error types depending
on the callsite.

Refactor the package to instead use its own ErrRepositoryNotFound error.
This also allows us to get rid of the `RepositoryNotFoundError` type,
where we instead inject error code and metadata in a central place in
the coordinator given that it has more context.