localrepo: Add `ConflictedFiles` function to `MergeTreeConflictError`
The error `MergeTreeConflictError`, provides information about merge
conflicts which have occurred during the usage of git-merge-tree(1).
Sometimes, we just want to know the conflicted files against the created
tree and don't care about other conflict information.

To make it easier to get the conflict tree directly, introduce
`ConflictedFiles` as a method on the error.