praefect: Wire `praefect.Router` to `NewGRPCServer()`
In a future commit, a Praefect stream handler will be registered for the
`GetObjectPool()` RPC. This handler will need to use `praefect.Router`
to handle routing the RPC to the appropriate node.

Make the necessary adjustments to wire `praefect.Router` to
`NewGRPCServer()` where the new stream handler will be registered in the
future.