.gitlab-ci.yml: Add rails job
Adding a job to run the full rails suite of tests. This allows
developers to trigger the job to see if their changes pass the rails
test suite.

We have three jobs to accomplish this. First, start-rails-specs will
create a branch and modify the GITALY_SERVER_VERSION on it. Then,
rails-specs will run the Rails test suite on that branch. Finally,
rails-specs-cleanup will delete that branch.