storage: Add option to skip storage check
In some cases, such as `CreateFork`, Praefect may not rewrite the
storage name to one present on the receiving Gitaly node. In these cases
we must skip validating that the storage is present on disk.

Add a `SkipStorageExistenceCheck` option to the storage locator to allow
us to skip this test when needed. This necessarily skips the repository
check as well, as we don't have a storage path to join to.