repository: No storage check for fork source repo
With 79e947645 (repository: Refactor repository validation to verify
existence, 2023-06-12) we started checking that the `SourceRepository`
storage name passed to `CreateFork` is valid and exists on the receiving
Gitaly node. Previously we confirmed this value was not empty, but
did not actually check that storage existed in the config or was present
on disk.

This field is unusual in that Praefect intentionally does not rewrite it
when proxying, which caused fork requests using the new validation
checks to fail with:

  validating source repository: storage name not found

Use the newly added `WithSkipStorageExistenceCheck` option to elide this
check.