Add expected_old_oid to UserMergeToRefRequest
This can be used to guard against races in composite git operations that
act as a single user-facing operation. The motivation here is the
construction of merge request refs using "semi-linear" merges with
automatic rebase prior to the creation of the merge commit.

See https://gitlab.com/gitlab-org/gitlab/-/issues/26996.

Changelog: changed