conflict: Move constants to new file `resolve.go`
The conflicts package defines two functions on the `File` structure,
i.e. `Parse` and `Resolve`. The two functions are tied to the `File`
structure which is tied to the Git2Go implementation.

In the upcoming commits we'll introduce a new standalone `Resolve`
function which is not tied to `File` structure, this allows us to
deprecate the `File` structure along with Git2Go:ResolveConflict
deprecation.

Let's move some of the variables which will be common to a new file
`resolve.go`, which would make it easier to deprecate the entire
`parser.go` file in the future.