conflict: Introduce new `Resolve` function
The conflicts package defines two functions on the `File` structure,
i.e. `Parse` and `Resolve`. The two functions are tied to the `File`
structure which is tied to the Git2Go implementation.

To make it easier to deprecate the old code and remove the dependency on
the `File` structure, we ignore DRY and copy over the `Parse` &
`Resolve` methods of the `File` structure to a standalone `Resolve`
function.

The new standalone function, takes in a conflicted blob as an
`io.Reader` along with the resolutions and then provides the resolved
blob as a new `io.Reader`. This makes it easier to use.

Eventually once we rollout `ResolveConflicts` RPC to use this
implementation, we can delete `parser[_test].go`.