gitaly: Log cause of the failed reference update
When Prepare() or Commit() method of the Updater was called inside
UpdaterWithHooks#UpdateReference method the errors returned were
thrown away. Because of it we were missing important part of the
failure cause - the output of the failed git operation. In order to
fix it now we return it as part of the Error instance. Because
returned error is a structured error with metadata assigned to it
this metadata will be logged by the interceptors around gRPC method.
The metadata is not returned to the RPC caller, so there are no changes
to the existing API.

Closes: #4677

Changelog: added