backup: Add vanity repository to restore request
The vanity repository allows using a different repository to calculate
repository paths from the repository that is being restored. This will
be used for praefect requests where the repository being restored gets
rewritten. So for praefect we accept the original, not rewritten
repository, such that backups are stored based on the praefect path, and
use the rewritten repository to access the git repository.