repository: Avoid NLP to detect license from README
We recently had an incident[1] having the go-license-detector eat up a
lot of resources. Research showed it was related to Natural Language
Processing was happening on the README in several repositories.

This step of license detection, also called "Plan B" in the
go-license-detector source code, is overkill and with this change we
skip it.

[1]: https://gitlab.com/gitlab-com/gl-infra/production/-/issues/15915