smarthttp: Fix PostReceivePack test behaviour with Praefect
Some of the tests for PostReceivePack assert whether Gitaly correctly
validates invalid requests. A subset of those tests has diverging
behaviour for Praefect though because it is being called with a repo
that does not exist in the database. This is not the behaviour that we
intend to test.

Fix these by creating the repository via `gittest.CreateRepository()` so
that Praefect knows about its existence.