repository: Fix CreateFork test behaviour with Praefect
Some of the tests for CreateFork assert whether Gitaly correctly
validates invalid requests. One of these tests is intended to check for
the case where the source repository wasn't provided. But because the
repository has already been created at that point, Praefect has
diverging behaviour from Gitaly as it will refuse to re-create the
repository.

Fix this by instead asking for a new repository to be created.