praefect/coordinator: Unify ErrRepositoryNotSet errors with Gitaly
The coordinator prefixes most of its errors with "repo scoped", which
makes the errors mismatch those returned by Gitaly. Consequentially,
Praefect does not act as an actually-transparent proxy. This can be
painfully felt in our tests, where we need to distinguish between both
errors.

Drop the prefix and return the error directly to unify the errors that
both Gitaly and Praefect would return.