git: Allow more pseudo-options for git-rev-list(1)
The git-rev-list(1) command is somewhat special in that it accepts
pseudo-options as regular arguments. That is, next to revisions like
`refs/heads/main`, it also accepts pseudo-options like `--all` or
`--branch`. These need to be intermixed as the order of arguments is
indeed important when pseudo-options like `--not` are passed in, as
well.

We have thus special-cased git-rev-list(1) to accept a subset of such
pseudo-options via arguments even though we normally reject positional
arguments which start with a dash. The list of accepted pseudo-options
is hardcoded so that we indeed only accept actual pseudo-options and no
actual options that would change behaviour of the command.

We have unified validation of these options recently in c1b8d6a09 (git:
Optionally allow pseudo-revisions in `ValidateRevision()`, 2023-06-05),
but forgot to also update this location. Convert it so that the list of
accepted pseudo-options needs to be maintained in a single location,
only.