remoterepo: Add HeadReference
HeadReference is a replacement for GetDefaultBranch. The intention
is to replace and remove the default branch heuristic that
GetDefaultBranch uses and instead rely solely on HEAD.

This could have been implemented as a parameter on GetDefaultBranch but
this requires modifying all call sites and localrepo already has a
method called HeadReference for this purpose.