coordinator: Remove prefix for missing repository errors
The coordinator uses its own hand-crafted error for when a repository
cannot be found when rewriting repository-scoped messages. Furthermore,
it uses an error prefix that makes the errors mismatch those returned by
Gitaly.

Drop the prefix and use `storage.NewRepositoryNotFoundError()` to unify
the errors that both Gitaly and Praefect would return.