ref: Fix FindRefsByOid test behaviour with Praefect
Some of the tests for FindRefsByOid assert whether Gitaly correctly
validates invalid requests. One of these tests checks whether the
request fails as expected when the repository is missing by first
creating the repository, but then deleting it on disk again. This causes
weird test behaviour on Praefect as the repository will exist in the
database, but doesn't on disk. So the actual observed error will be
returned by Gitaly and not by Praefect.

Refactor the test to just not create the repository in the first place,
which is equivalent to a missing repository. While we could instead also
call the `RemoveRepository()` RPC, this would only cause us to evaluate
whether that RPC correctly removes the repository as expected.