Sync repository creations to disk
Repository creations are currently not flushed to disk. This can lead
to data loss if the host crashes before the writes have been persisted
to disks. Sync the writes the the disk prior to acknowleding a repository
creations to avoid this. As we don't know whether the directory hierarchy
in the storage existed, we sync them always regardless, and assume that
the root storage directory's entry in the parent directroy is synced
already.

Changelog: fixed