Sync repository removals
Repository removals are currently not synced to the disk. This may
lead to the repository appearing back on the disk either in full or
in corrupted state with deletion partially applied. As we are moving
the repository to a temporary directory before deleting it, we need
to just sync the move the removal from the parent directory in order
to avoid the issues.

Changelog: fixed