Sync migrated quarantine directory contents
Quarantine directory's contents are not synced currently when they
get migrated to the repository's object database. This may lead to
some objects missing after a host crash. This commit adds the missing
syncs for the migrated objects and directories.

There's still a chance of repository corruption as the objects are not
migrated in a dependency order. We may successfully migrate a commit
into the target repository without its tree for example. This should be
handled separately.

Changelog: fixed