repository: Make WriteRef compatible with the SHA256 object format
The WriteRef RPC needs to check for the zero object ID. This currently
happens via a hardcoded call to the SHA1 object hash, which of course
breaks with SHA256 repositories.

Fix this by detecting the object hash at runtime so that the RPC becomes
compatible with SHA256 repositories and enable testing with this object
format.

Changelog: added