Implement a request constructor in protoregistry
The `Registry` type currently has `UnmarshalRequestProto` method which
allows for unmarshaling a given RPC's serialized request into the
correcty request type. In some contexts, like with when receiving
messages with `grpc.ServerStream.RecvMsg()`, we don't have the the
marshaled request available but have to provide an instance of the
request type where the request gets unmarshaled to. To make generically
possible, this commit implements `MethodInfo.NewRequest()` which just
returns a new instance of the request type. This instance can then be
passed to the `RecvMsg` for unmarshaling the request.

The implementation around `UnmarshalRequestProto` was also simplified by
using the new method.