Implement helper for conditional testing with WAL
We are soon plugging in the transaction management and running all
tests with it enabled. While there mostly shouldn't be any behavioral
differences, there may be some cases where the tests are testing
scenarios that shouldn't really happen with transactions or asserting
specific error messages. This commit introduces a helper that can
be used to conditionally change the test behavior depending on whether
or not transaction management is enabled.