Add SkipWithWAL to conditionally skip tests with WAL
Some tests are testing scenarios which which are not really meaningful
to test with WAL enabled. Some examples include testing with symlinks
in the repository where we generally wouldn't have symlinks in production
nor would the WAL logic ever write them. The helper can be used to skip
such test but keep the tests still in place until the migration is
complete and they can be removed.