Do not assert disk path in ObjectFormat tests
ObjectFormat tests is asserting that the exact disk path is returned
in the error metadata. We should not assert exact disk paths when testing
through the API as they may change and force us to touch all tests asserting
the disk paths. This commit changes the test to assert the remainder of the
metadata value but not the path itself. This is achieved with the new helper
function introduced for this use case.