Don't assert repository path in Fsck test
Fsck tests contain a test case that is asserting the repository's
path. This is not desired as the repository paths are an internal
detail that may change. The path changing would require us to touch
up every test that is asserting it. This commit changes the test case
to assert the remainder of the message but not the repository path.